Added endpoint for checking duplicate. Refactored Kabin integration classes.
