--- src/main/kotlin/no/nav/klage/oppgave/api/view/GetCompletedKlagebehandlingerInput.kt
@@ -1,5 +1,0 @@
-package no.nav.klage.oppgave.api.view
-
-data class GetCompletedKlagebehandlingerInput(
-    val idnummer: String
-)
