--- src/main/kotlin/no/nav/klage/oppgave/api/view/SearchUsedJournalpostIdInput.kt
@@ -1,5 +1,0 @@
-package no.nav.klage.oppgave.api.view
-
-data class SearchUsedJournalpostIdInput(
-    val fnr: String,
-)
